Refactor ResumeSection and ResumeFeature into new file
